Minor fixes with new compilation flags

+++ creational_patterns/builder_pattern/builder_example_with_html/builder_pattern_example_with_html_code_creation.cpp
@@ -44,9 +44,10 @@ struct HtmlElement {
   // API for printing elements
   std::string str(int indent = 0) const {
     std::ostringstream oss{};
-    std::string i(indent_size * indent, ' ');
+    std::string i(indent_size * static_cast<long unsigned int>(indent), ' ');
     oss << i << "<" << name << ">" << std::endl;
-    if (text.size() > 0) oss << std::string(indent_size * (indent + 1), ' ') << text << std::endl;
+    if (text.size() > 0)
+      oss << std::string(indent_size * static_cast<long unsigned int>(indent + 1), ' ') << text << std::endl;
 
     for (const auto& e : elements) oss << e.str(indent + 1);
